homework_6/my_site/pyproject.toml
[tool.poetry]
name = "my_site"
version = "0.1.0"
description = ""
authors = ["Pavel <giropavel85@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
Psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
